Migrating the map page to seperate file
