+++ lib/main.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: prefer_const_constructors
 
 import 'dart:async';
+import 'package:bike_tour_app/screens/navigation/main_map.dart';
 import 'package:bike_tour_app/screens/authenticate/authenticate.dart';
 import 'package:bike_tour_app/screens/wrapper.dart';
 import 'package:firebase_auth/firebase_auth.dart';
@@ -37,56 +38,6 @@ class MyApp extends StatelessWidget {
       routes: {
         '/': (context) => Authenticate(),
       },
-    );
-  }
-}
-
-class MyApp1 extends StatefulWidget {
-  const MyApp1({Key? key}) : super(key: key);
-
-  @override
-  _MyAppState1 createState() => _MyAppState1();
-}
-
-class _MyAppState1 extends State<MyApp1> {
-  late GoogleMapController mapController;
-
-  final LatLng _center = const LatLng(51.507399, -0.127689);
-
-  void _onMapCreated(GoogleMapController controller) {
-    mapController = controller;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-          body: Center(
-            child: GoogleMap(
-              onMapCreated: _onMapCreated,
-              initialCameraPosition: CameraPosition(target: _center, zoom: 15),
-            ),
-          ),
-          floatingActionButton: Stack(
-            children: <Widget>[
-              Align(
-                alignment: Alignment(1, -0.8),
-                child: FloatingActionButton(
-                  onPressed: () {},
-                  backgroundColor: Color.fromARGB(202, 85, 190, 56),
-                  child: const Icon(Icons.settings),
-                ),
-              ),
-              Align(
-                alignment: Alignment(-0.8, -0.8),
-                child: FloatingActionButton(
-                  onPressed: () {},
-                  backgroundColor: Color.fromARGB(202, 85, 190, 56),
-                  child: const Icon(Icons.person),
-                ),
-              ),
-            ],
-          )),
     );
   }
 }
